feat: linux only code from attrs moved to a linux only file
