+++ src/fuse/fuse_impl.rs
@@ -1,4 +1,4 @@
-use crate::fuse::linux_attr::time_or_now_to_system_time;
+use crate::fuse::linux_attrs::time_or_now_to_system_time;
 use crate::pods::arbo::{FsEntry, Inode};
 use crate::pods::filesystem::attrs::SetAttrError;
 use crate::pods::filesystem::fs_interface::{FsInterface, SimpleFileType};
