+++ src/fuse/mod.rs
@@ -1,2 +1,2 @@
 pub mod fuse_impl;
-pub mod linux_attr;
+pub mod linux_attrs;
